fastapi-rag/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastapi-rag"
version = "0.1.0"
description = "FastAPI RAG system for Miguel's portfolio"
authors = [
    {name = "Miguel Parracho", email = "miguel@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "httpx==0.27.0",
    "python-dotenv==1.0.1",
    "asyncpg==0.29.0",
    "pgvector==0.2.5",
    "markdown==3.6",
    "openai==1.40.0"
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.0",
    "pytest-asyncio==0.24.0",
    "black==24.1.1",
    "isort==5.13.2",
    "flake8==7.0.0"
]

[project.scripts]
fastapi-rag = "app.main:main"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
